added jwt auth for apis

+++ backend/api/v1/views/resources.py
@@ -1,7 +1,8 @@
-from flask import Blueprint, jsonify, request
+from flask import Blueprint, jsonify, request, abort
+from flask_jwt_extended import jwt_required, current_user
 from app import db
 
-from models.Resource import Resource
+from models import Resource, Role, User
 
 
 bp = Blueprint("resources", __name__, url_prefix='/resource')
@@ -16,11 +17,23 @@ def get_resources():
 @bp.route("/<int:id>", methods=['GET', 'PUT', 'DELETE'])
 def get_resource(id):
     resource = Resource.query.filter(Resource.id == id).first_or_404()
+    return resource.to_dict()
+
+
+@bp.route("/<int:id>", methods=['PUT', 'DELETE'])
+@jwt_required
+def update_delete_resource(id):
+    resource = Resource.query.filter(Resource.id == id).first_or_404()
+    if current_user.id != resource.user_id or is_admin(current_user.id):
+        abort(400)
     if request.method == 'PUT':
         args = request.get_json()
         resource.update(**args)
-        db.session.commit()
     elif request.method == 'DELETE':
         db.session.delete(resource)
-        db.session.commit()
-    return resource.to_dict()
+
+
+def is_admin(id):
+    admin = Role.query.filter(Role.name == 'admin')
+    user = User.query.filter(User.id == id)
+    return user.role == admin
